Move core into subfolders
