+++ docs/pages/guide/demo.tsx
@@ -5,8 +5,7 @@ import { Field } from '@spark-web/field';
 import { Heading } from '@spark-web/heading';
 import { Hidden } from '@spark-web/hidden';
 import { CheckCircleIcon } from '@spark-web/icon';
-import { Inline } from '@spark-web/layout';
-import { Columns, Stack } from '@spark-web/layout';
+import { Columns, Inline , Stack } from '@spark-web/layout';
 import { Text } from '@spark-web/text';
 import { TextInput } from '@spark-web/text-input';
 import { TextLink } from '@spark-web/text-link';
