+++ packages/checkbox/src/Checkbox.stories.tsx
@@ -1,7 +1,6 @@
 import { Field } from '@spark-web/field';
 import { InformationCircleIcon } from '@spark-web/icon';
-import { Inline } from '@spark-web/layout';
-import { Stack } from '@spark-web/layout';
+import { Inline , Stack } from '@spark-web/layout';
 import { Text } from '@spark-web/text';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
