+++ packages/core/src/a11y/emoji.tsx
