+++ packages/core/src/a11y/index.ts
@@ -1,0 +1,9 @@
+export { Emoji } from './emoji';
+export { useFocusRing, useFocusVisible } from './focus-ring';
+export { composeId, IdProvider, mergeIds, useId } from './id-context';
+export { VisuallyHidden } from './visually-hidden';
+
+// types
+
+export type { EmojiProps } from './emoji';
+export type { BackgroundTone } from './focus-ring';
