+++ packages/core/src/index.ts
@@ -1,43 +1,58 @@
-export { useBackground, useBackgroundLightness } from './background-context';
-export { BackgroundProvider } from './background-provider';
-export { Box } from './box';
-export { buildDataAttributes } from './build-data-attributes';
-export { Emoji } from './emoji';
-export { useFocusRing, useFocusVisible } from './focus-ring';
+export {
+  composeId,
+  Emoji,
+  IdProvider,
+  mergeIds,
+  useFocusRing,
+  useFocusVisible,
+  useId,
+  VisuallyHidden,
+} from './a11y';
+export {
+  BackgroundProvider,
+  Box,
+  useBackground,
+  useBackgroundLightness,
+} from './box';
 export {
   AestheticoStylesheet,
   FontStylesheets,
   InterStylesheet,
 } from './fonts';
-export { forwardRefWithAs } from './forward-ref-with-as';
-export { composeId, IdProvider, mergeIds, useId } from './id-context';
 export { makeLinkComponent, useLinkComponent } from './link-context';
-export { makeBrighteTheme } from './make-theme';
-export { mergeRefs } from './merge-refs';
-export { resetElementStyles } from './reset-element-styles';
 export { SparkProvider } from './spark-provider';
-export { defaultTheme, ThemeProvider, useTheme } from './theme-context';
-export { createResponsiveMapFn } from './theme-utils';
-export { defaultTokens } from './tokens';
-export { isBoolean, isFunction, isNumber, isString } from './type-check';
-export { assignRef, useComposedRefs } from './use-composed-refs';
-export { useDisclosure } from './use-disclosure';
-export { typedKeys } from './utility-functions';
-export { VisuallyHidden } from './visually-hidden';
+export {
+  createResponsiveMapFn,
+  defaultTheme,
+  defaultTokens,
+  makeBrighteTheme,
+  ThemeProvider,
+  useTheme,
+} from './theme';
+export { buildDataAttributes } from './utils/build-data-attributes';
+export { forwardRefWithAs } from './utils/forward-ref-with-as';
+export { mergeRefs } from './utils/merge-refs';
+export { resetElementStyles } from './utils/reset-element-styles';
+export { isBoolean, isFunction, isNumber, isString } from './utils/type-check';
+export { assignRef, useComposedRefs } from './utils/use-composed-refs';
+export { useDisclosure } from './utils/use-disclosure';
+export { typedKeys } from './utils/utility-functions';
 
 // types
 
-export type { BackgroundProviderProps } from './background-provider';
-export type { BoxProps } from './box';
-export type { DataAttributeMap } from './build-data-attributes';
-export type { EmojiProps } from './emoji';
-export type { BackgroundTone } from './focus-ring';
+export type { BackgroundTone, EmojiProps } from './a11y';
+export type { BackgroundProviderProps, BoxProps } from './box';
 export type { LinkComponent, LinkComponentProps } from './link-context';
-export type { BrighteTextDefinition, BrighteTheme } from './make-theme';
 export type { SparkProviderProps } from './spark-provider';
-export type { ResponsiveProp, ResponsiveRangeProps } from './theme-utils';
+export type {
+  BrighteTextDefinition,
+  BrighteTheme,
+  ResponsiveProp,
+  ResponsiveRangeProps,
+} from './theme';
+export type { DataAttributeMap } from './utils/build-data-attributes';
 export type {
   AssignableRef,
   ElementTagNameMap,
   SingleOrArray,
-} from './utility-types';
+} from './utils/utility-types';
