+++ packages/core/src/theme/index.ts
@@ -1,0 +1,9 @@
+export { makeBrighteTheme } from './make-theme';
+export { defaultTheme, ThemeProvider, useTheme } from './theme-context';
+export { createResponsiveMapFn } from './theme-utils';
+export { defaultTokens } from './tokens';
+
+// types
+
+export type { BrighteTextDefinition, BrighteTheme } from './make-theme';
+export type { ResponsiveProp, ResponsiveRangeProps } from './theme-utils';
