+++ packages/core/src/utils/build-data-attributes.ts
