+++ packages/core/src/utils/merge-refs.ts
