+++ packages/core/src/utils/utility-functions.ts
